Fix - Bug route / Add - fetch update connected / Add - avatar

+++ model/User.js
@@ -6,7 +6,7 @@ const myPlaintextPassword = "145OkyayNo668Pass";
 const FILE_PATH = __dirname + "./../data/users.json";
 
 class User {
-  constructor(username, email, password, fName, lName, idUser) {
+  constructor(username, email, password, fName, lName, idUser, connected) {
     //TODO modifié et gérer l'utilisateur NOUVEAU et Existant
     this.idUser = idUser;
     //this.idUser = getUserListFromFile(FILE_PATH).length+1;
@@ -18,7 +18,9 @@ class User {
     this.avatar = null;
     this.type = "users";
     this.itemCollections = [];
+    this.connected = connected;
   }
+
 
   /* return a promise with async / await */ 
   async save() {
@@ -35,6 +37,7 @@ class User {
       avatar: this.avatar,
       type: this.type,
       itemCollections: this.itemCollections,
+      connected: this.connected
     });
     saveUserListToFile(FILE_PATH, userList);
     return true;
@@ -44,10 +47,9 @@ class User {
   checkCredentials(username, password) {
     if (!username || !password) return false;
     let userFound = User.getUserFromList(username);
-    console.log("User::checkCredentials:", userFound, " password:", password);
+    //console.log("User::checkCredentials:", userFound, " password:", password);
     if (!userFound) return Promise.resolve(false);
-    //try {
-    console.log("checkCredentials:prior to await");
+    //console.log("checkCredentials:prior to await");
     // return the promise
     return bcrypt
       .compare(password, userFound.password)
@@ -55,31 +57,11 @@ class User {
       .catch((err) => err);
   }
 
-  // Some example of bcrypt used with sync function
-  /*
-  save() {
-    let userList = getUserListFromFile(FILE_PATH);
-    const hashedPassword = bcrypt.hashSync(this.password, saltRounds);
 
-    userList.push({
-      username: this.email,
-      email: this.email,
-      password: hashedPassword,
-    });
-
-    saveUserListToFile(FILE_PATH, userList);
+  static updateConnection(value, userId){
+    updateConnectedField(value, userId, FILE_PATH);
   }
-
-  checkCredentials(email, password) {
-    if (!email || !password) return false;
-    let userFound = User.getUserFromList(email);
-    console.log("User::checkCredentials:", userFound, " password:", password);
-    if (!userFound) return false;
-    const match = bcrypt.compareSync(password, userFound.password);
-    return match;
-  }*/
-
-  static get list() {
+  static getList() {
     let userList = getUserListFromFile(FILE_PATH);
     return userList;
   }
@@ -140,6 +122,25 @@ function getUserListFromFile(filePath) {
 
 function saveUserListToFile(filePath, userList) {
   const fs = require("fs");
+  let data = JSON.stringify(userList);
+  fs.writeFileSync(filePath, data);
+}
+function updateConnectedField(value, userId, filePath){
+  const fs = require("fs");
+  if (!fs.existsSync(filePath)) return [];
+  let userListRawData = fs.readFileSync(filePath);
+  let userList;
+  if (userListRawData) {
+
+    userList = JSON.parse(userListRawData);
+  }
+
+  for(let i = 0; i < userList.length; i++){
+    if(userList[i].idUser == userId ){
+      userList[i].connected = value;
+      break;
+    }
+  }
   let data = JSON.stringify(userList);
   fs.writeFileSync(filePath, data);
 }
